Adicionando componentes necessarios a pagine de podcast, além de componentizar a rendenização dos cards de desktop emmobile e corrigindo tamanho de container nas paginas.
