+++ src/components/BackGroundImg/BackGroundImg.module.css
@@ -67,13 +67,13 @@
 }
 
 @media(min-width: 1440px) {
-    .titleContainer {
+    .container {
         height: 550px;      
     }
 }
 
 @media(min-width: 1920px) {
-    .titleContainer {
+    .container {
         width: 80%;      
     }
 }
