+++ src/components/CardMobile/index.jsx
@@ -2,26 +2,26 @@ import styles from "./CardMobile.module.css"
 
 import img1 from "../../assets/invocacao-do-mal-4.jpg"
 
-export function CardMobile() {
+export function CardMobile({ tag, title, author }) {
     return (
         <div className={styles.container}>
             <div className={styles.imgElement}>
-                <img src={img1} alt="" />
+                <img src={img1} alt="Img" />
 
                 <div className={styles.tag}>
-                    <p>Cinema</p>
+                    <p>{tag}</p>
                 </div>
             </div>
 
             <div className={styles.infoContainer}>
                 <h3 className={styles.title}>
-                    O vale da estranheza presente em mad god.
+                    {title}
                 </h3>
                 
 
                 <div className={styles.author}>
                     <p className={styles.firstPart}>por</p>
-                    <p className={styles.secondPart}>Gabriel Zanon</p>
+                    <p className={styles.secondPart}>{author}</p>
                 </div>
             </div>
