+++ src/components/CardRender/CardRender.module.css
@@ -1,0 +1,5 @@
+.cardsContainer {
+    display: flex;
+    flex-direction: column;
+    gap: 25px;
+}
